[music] Add placeholder notice
